[14/09/2024] Forum Backup

+++ qrCodeGen.php
@@ -36,51 +36,65 @@
         <br>
         <br>
         <h1>PDF Sendo Gerado</h1>
+        <br>
+        <br>
+        <h1 id="txtDsgn">Gerando</h1>
         <script src="js/qrCodeGen.js"></script>
             <?php
                 require "conexao.php";
                 
-                $sqlChecker = "SELECT * FROM tb_pessoa ORDER BY nm_pessoa ASC;"; // verificador se a pessoa já cadastrou ou não
+                $sqlChecker = "SELECT nm_pessoa, id_pessoa, cpf_pessoa FROM tb_pessoa ORDER BY nm_pessoa ASC;"; // verificador se a pessoa já cadastrou ou não
                 $resultCheck = $conn->query($sqlChecker);
 
-                $iEtiqueta = 0;
-                $pag = 1;
-                $iCaseiro = 1;
-                $iPrimo = 1; // não sendo primo!
+                $iEtiqueta = 0; // numero da etiqueta até o momento em 1 página
+                $pag = 1;       // página atual
+                $iCaseiro = 1;  // (de caseiro tem nada!) normalmente utilizado como posição Y no script JS
+                $iPrimo = 1;    // (não sendo primo!) normalmente utilizado como posição X no script JS
 
                 while($row = $resultCheck->fetch_assoc()) { // todos os resultados demonstrados do curso 
-                    $sqlChecker2 = "SELECT * FROM tb_cadastrado WHERE fk_cpf_pessoa = '".$row['cpf_pessoa']."';"; // verificador se a pessoa já cadastrou ou não
-                    $resultCheck2 = $conn->query($sqlChecker2);
-                    echo '<script>down(\''.$row['nm_pessoa'].'\',
-                    \''.$row['id_pessoa'].'\',
-                    \'https://forumetecab.com.br/senha.php?cpf='.$row['cpf_pessoa'].'\',
-                    \''.$iPrimo.'\',
-                    \''.$iCaseiro.'\',
-                    \''.$pag.'\'
-                    )</script>';
-                    if ($iPrimo == 3){
-                        $iPrimo = 1;
-                        $iCaseiro = $iCaseiro + 1;
-                    }
-                    else{
-                        $iPrimo = $iPrimo + 1;
-                    }
-                    $iEtiqueta = $iEtiqueta + 1;
-                    if ($iEtiqueta >= 33){
-                        $iEtiqueta = 0;
-                        $pag = $pag + 1;
-                        $iPrimo = 1;
-                        $iCaseiro = 1;
+                    $sqlChecker3 = "SELECT * FROM tb_imprimido WHERE fk_cpf_pessoa = '".$row['cpf_pessoa']."';"; // verificador se a pessoa já cadastrou ou não
+                    $resultCheck3 = $conn->query($sqlChecker3);
+                    if ($resultCheck3->num_rows <= 0){ // se não existir o login/cadastro
+                        echo '<script>down(\''.$row['nm_pessoa'].'\',
+                        \''.$row['id_pessoa'].'\',
+                        \'https://forumetecab.com.br/senha.php?cpf='.$row['cpf_pessoa'].'\',
+                        \''.$iPrimo.'\',
+                        \''.$iCaseiro.'\',
+                        \''.$pag.'\'
+                        );
+                        </script>';
+                        if ($iPrimo == 3){ // o '3' significa a quantidade de colunas de etiquetas
+                            $iPrimo = 1;
+                            $iCaseiro = $iCaseiro + 1;
+                        }
+                        else{
+                            $iPrimo = $iPrimo + 1;
+                        }
+                        $iEtiqueta = $iEtiqueta + 1;
+                        if ($iEtiqueta >= 33){  // quantidade de etiquetas em 1 página
+                            $iEtiqueta = 0;
+                            $pag = $pag + 1;
+                            $iPrimo = 1;
+                            $iCaseiro = 1;
+                        }
                     }
                 }
-                echo '<script>save();</script>';
+                echo '
+                <script>document.getElementById("txtDsgn").setAttribute("hidden","true")</script>
+                <h1>Gerado!</h1>
+                <br>
+                <br>
+                ';
             ?>
+            <button class="btn" onclick="save()">Imprimir</button>
+            <br>
+            <br>
+            <button class="btn" onclick="window.location.href='registraEtiqueta.php?&g-logged=GLOG'">Registrar Cadastros</button>
     </div>
 
     <div class="footer" style="margin-top: 25%;">
         <h5>Site desenvolvido pelos alunos
             <br><a href="https://github.com/niButera">Nicolas</a> e <a href="https://github.com/yukio-sato">Yukio</a>
-            3i3 - 1º Semestre - 2024
         </h5>
     </div>
 </body>
